server/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-storage
google-auth
pydantic
scikit-learn
pandas
fastparquet
pyarrow